+++ tests/Unit/SampleTestWithReallyLongFilenameForNotFormattedTest.php
@@ -1,26 +1,12 @@
 <?php
 
-use Codedungeon\PHPUnitPrettyResultPrinter\Printer;
+declare(strict_types=1);
 
-/**
- *  Corresponding Class to test Printer class.
- *
- * @author mike erickson
- */
-class SampleTestWithReallyLongFilenameForNotFormattedTest extends PHPUnit\Framework\TestCase
+use PHPUnit\Framework\TestCase;
+
+final class SampleTestWithReallyLongFilenameForNotFormattedTest extends TestCase
 {
-    /**
-     * @var
-     */
-    protected $printer;
-
-    public function setUp(): void
-    {
-        $this->printer = new Printer();
-    }
-
-    /** @test  */
-    public function it_should_pass_test_part_of_long_filename(): void
+    public function testItPassesAsVisualCheckForLongClassName(): void
     {
         $this->assertTrue(true);
     }